+++ odoo_health_check/models/__init__.py
@@ -1,3 +1,4 @@
 from . import ir_cron_history
 from . import ir_cron
+from . import health_check_result
 from . import res_config_settings