+++ odoo_health_check/tests/__init__.py
@@ -2,3 +2,4 @@ from . import test_history_model
 from . import test_cron_override
 from . import test_failure_email
 from . import test_config
+from . import test_disk_check